requirements.txt
openai
azure-functions
azure-storage-queue
